Refactor UserWithReposRepository to IndexedDataSource

+++ app/src/main/java/com/husseinelfeky/githubpaging/api/GitHubApi.kt
@@ -9,15 +9,9 @@ import retrofit2.http.Query
 
 interface GitHubApi {
 
-    /**
-     * GitHub Users API endpoint does not support pagination by page; it paginates
-     * with respect to the user's id.
-     *
-     * @param sinceId Set it to "(pageToFetch - 1) * pageSize" to solve this issue.
-     */
     @GET("users")
     fun getUsers(
-        @Query("since") sinceId: Int,
+        @Query("since") sinceId: Long,
         @Query("per_page") pageSize: Int
     ): Single<List<User>>
