+++ app/src/main/java/com/husseinelfeky/githubpaging/persistence/dao/GitHubDao.kt
@@ -13,10 +13,10 @@ interface GitHubDao {
 
     @Transaction
     @Query("SELECT * FROM users ORDER BY id LIMIT :itemsLimit")
-    fun getUsersWithRepos(itemsLimit: Int): Flowable<List<UserWithRepos>>
+    fun getUsersWithRepos(itemsLimit: Long): Flowable<List<UserWithRepos>>
 
     @Query("SELECT * FROM users ORDER BY id LIMIT :pageSize OFFSET :offset")
-    fun getUsers(pageSize: Int, offset: Int): Single<List<User>>
+    fun getUsers(pageSize: Int, offset: Long): Single<List<User>>
 
     @Query("SELECT * FROM repositories WHERE userId = :userId ORDER BY name")
     fun getUserRepos(userId: Long): Single<List<GitHubRepo>>
